add phase as InterpolatePoints
