+++ qiskit_pasqal_provider/providers/gate.py
@@ -2,142 +2,20 @@
 
 from typing import Any, Union
 
-import numpy as np
 from numpy.typing import ArrayLike
 
 from pulser.math import AbstractArray
-from qiskit.circuit import Parameter, ParameterExpression
+from qiskit.circuit import ParameterExpression
 from qiskit.circuit.gate import Gate
 
 from qiskit_pasqal_provider.providers.pulse_utils import (
     PasqalRegister,
     RegisterTransform,
     GridLiteralType,
+    InterpolatePoints,
 )
 
 CoordsKey = Union[str, int, float]
-
-
-class InterpolatePoints:
-    """
-    A class to hold attributes for later use on `pulser`'s `InterpolateWaveform` class.
-    It should be used to generate the points for the `HamiltonianGate` instance.
-    """
-
-    __slots__ = (
-        "_values",
-        "_duration",
-        "_times",
-        "_n",
-        "_params",
-        "_interpolator",
-        "_interpolator_kwargs",
-    )
-
-    def __init__(
-        self,
-        values: ArrayLike | ParameterExpression,
-        duration: int | float | ParameterExpression = 1000,
-        times: ArrayLike | None = None,
-        n: int | None = None,
-        interpolator: str = "PchipInterpolator",
-        **interpolator_kwargs: Any,
-    ):
-        """
-        A class to hold attributes for later use on `pulser`'s `InterpolateWaveform`
-        class. It should be used to generate the points for the `HamiltonianGate`
-        instance.
-
-        Args:
-            values: an array-like data representing the points to be interpolated.
-                It can be parametrized through `qiskit.circuit.Parameter`.
-            duration: optional duration of the waveform (in ns). Defaults to 1000.
-            times: Fractions of the total duration (between 0 and 1). Optional.
-            n: the number of values points, in case `qiskit.circuit.Parameter` is
-                provided on values argument. Default to None.
-            interpolator: The SciPy interpolation class to use. Supports
-                "PchipInterpolator" and "interp1d".
-            **interpolator_kwargs: Extra parameters to give to the chosen
-                interpolator class.
-        """
-
-        assert isinstance(duration, int | float | ParameterExpression)
-        assert isinstance(interpolator, str)
-
-        if n is None:
-            values = np.array(values)
-            n = len(values)
-
-        elif isinstance(values, ParameterExpression) and None is not n:
-            values = np.full(shape=n, fill_value=values)
-
-        else:
-            raise ValueError("Argument 'n' must be the size of values argument.")
-
-        self._n = n
-        self._values = values
-        self._duration = duration
-        self._times = times
-        self._params = self._extract_params()
-        self._interpolator = interpolator
-        self._interpolator_kwargs = interpolator_kwargs
-
-    @property
-    def duration(self) -> int | float | Parameter:
-        """duration of the waveform (in ns)"""
-        return self._duration
-
-    @property
-    def values(self) -> ArrayLike:
-        """data points for interpolation"""
-        return self._values
-
-    @property
-    def times(self) -> ArrayLike | None:
-        """normalized fraction of the total duration. Can be None"""
-        return self._times
-
-    @property
-    def parameters(self) -> list[Parameter | ParameterExpression]:
-        """list of parameters"""
-        return self._params
-
-    @property
-    def interpolator(self) -> str:
-        """The interpolator method name."""
-        return self._interpolator
-
-    @property
-    def interpolator_options(self) -> dict:
-        """The key-value pairs to fill the interpolator function with."""
-        return self._interpolator_kwargs
-
-    def _extract_params(self) -> list[Parameter | ParameterExpression]:
-        """Extract the parameters list from values, duration and times arguments."""
-
-        values_params = []
-        for k in self.values:  # type: ignore [union-attr]
-            if isinstance(k, Parameter | ParameterExpression):
-                values_params.extend(k.parameters)
-
-        duration_params = (
-            list(self.duration.parameters)
-            if isinstance(self.duration, ParameterExpression)
-            else []
-        )
-
-        times_params = []
-
-        if self.times is not None:
-            for k in self.times:  # type: ignore [union-attr]
-                if isinstance(k, Parameter | ParameterExpression):
-                    times_params.extend(k.parameters)
-
-        return list(set(values_params + duration_params + times_params))
-
-    def __len__(self) -> int:
-        """InterpolatePoints length is equal to its values' length."""
-        return len(self.values)
 
 
 class HamiltonianGate(Gate):
@@ -176,10 +54,14 @@ class HamiltonianGate(Gate):
             raise NotImplementedError("'composed_wf' argument is not available yet.")
 
         if not isinstance(amplitude, InterpolatePoints):
-            raise TypeError(f"amplitude must be InterpolatePoints, not {type(amplitude)}.")
+            raise TypeError(
+                f"amplitude must be InterpolatePoints, not {type(amplitude)}."
+            )
 
         if not isinstance(detuning, InterpolatePoints):
-            raise TypeError(f"detuning must be InterpolatePoints, not {type(detuning)}.")
+            raise TypeError(
+                f"detuning must be InterpolatePoints, not {type(detuning)}."
+            )
 
         if not isinstance(phase, InterpolatePoints | float | ParameterExpression):
             raise TypeError(
